pyproject.toml
[tool.poetry]
name = "data_project"
version = "0.1.0"
description = ""
authors = ["Olesya Zybareva"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.5.0"
tortoise-orm = "^0.19.2"
openpyxl = "^3.0.10"
aerich = "^0.7.1"
asyncpg = "^0.26.0"

[tool.aerich]
tortoise_orm = "src.model_description.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

